# openapi.json
{"openapi":"3.1.0","info":{"title":"API de Construcción","description":"Documentación de la API para la gestión de proyectos y equipos.","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/registro-equipos/equipos":{"post":{"summary":"Crear Equipo","operationId":"crear_equipo_registro_equipos_equipos_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipo"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"Listar Equipos","operationId":"listar_equipos_registro_equipos_equipos_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Equipo"},"title":"Response Listar Equipos Registro Equipos Equipos Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/registro-equipos/equipos/{equipo_id}":{"get":{"summary":"Obtener Equipo","operationId":"obtener_equipo_registro_equipos_equipos__equipo_id__get","parameters":[{"name":"equipo_id","in":"path","required":true,"schema":{"type":"string","title":"Equipo Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Actualizar Equipo","operationId":"actualizar_equipo_registro_equipos_equipos__equipo_id__put","parameters":[{"name":"equipo_id","in":"path","required":true,"schema":{"type":"string","title":"Equipo Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipo"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Equipo"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Eliminar Equipo","operationId":"eliminar_equipo_registro_equipos_equipos__equipo_id__delete","parameters":[{"name":"equipo_id","in":"path","required":true,"schema":{"type":"string","title":"Equipo Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/proyectos":{"get":{"summary":"Listar Proyectos","operationId":"listar_proyectos_proyectos_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Proyecto"},"type":"array","title":"Response Listar Proyectos Proyectos Get"}}}}}},"post":{"summary":"Crear Proyecto","operationId":"crear_proyecto_proyectos_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Proyecto"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Proyecto"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/proyectos/{proyecto_id}":{"get":{"summary":"Obtener Proyecto","operationId":"obtener_proyecto_proyectos__proyecto_id__get","parameters":[{"name":"proyecto_id","in":"path","required":true,"schema":{"type":"string","title":"Proyecto Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Proyecto"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"summary":"Actualizar Proyecto","operationId":"actualizar_proyecto_proyectos__proyecto_id__put","parameters":[{"name":"proyecto_id","in":"path","required":true,"schema":{"type":"string","title":"Proyecto Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Proyecto"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Proyecto"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Eliminar Proyecto","operationId":"eliminar_proyecto_proyectos__proyecto_id__delete","parameters":[{"name":"proyecto_id","in":"path","required":true,"schema":{"type":"string","title":"Proyecto Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/login":{"post":{"summary":"Login","operationId":"login_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/clientes/":{"post":{"summary":"Crear Cliente","operationId":"crear_cliente_clientes__post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Cliente"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/inventarios/registro":{"post":{"summary":"Registrar Inventario","operationId":"registrar_inventario_inventarios_registro_post","requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Data"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/cotizaciones/":{"post":{"summary":"Crear Cotizacion","operationId":"crear_cotizacion_cotizaciones__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Cotizacion"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/zona-restringida":{"get":{"summary":"Zona Restringida","operationId":"zona_restringida_admin_zona_restringida_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard":{"get":{"summary":"Dashboard","operationId":"dashboard_dashboard_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Cotizacion":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id","description":"ID único de la cotización"},"cliente_id":{"type":"string","title":"Cliente Id","description":"ID del cliente asociado a la cotización"},"total":{"type":"number","title":"Total","description":"Monto total de la cotización"}},"type":"object","required":["cliente_id","total"],"title":"Cotizacion"},"Equipo":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id","description":"ID único del equipo"},"nombre":{"type":"string","title":"Nombre","description":"Nombre del equipo"},"estado":{"type":"string","title":"Estado","description":"Estado actual del equipo (activo/inactivo)"},"ubicacion":{"type":"string","title":"Ubicacion","description":"Ubicación del equipo"},"materiales":{"items":{"type":"string"},"type":"array","title":"Materiales","description":"Lista de materiales asociados al equipo","default":[]},"proyectos":{"items":{"type":"string"},"type":"array","title":"Proyectos","description":"Lista de IDs de proyectos asociados al equipo","default":[]},"fecha_asignacion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fecha Asignacion","description":"Fecha de asignación del equipo a un proyecto"},"fecha_desasignacion":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fecha Desasignacion","description":"Fecha de desasignación del equipo de un proyecto"}},"type":"object","required":["nombre","estado","ubicacion"],"title":"Equipo"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"usuario":{"type":"string","title":"Usuario"},"password":{"type":"string","title":"Password"}},"type":"object","required":["usuario","password"],"title":"LoginRequest"},"Proyecto":{"properties":{"id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Id","description":"ID único del proyecto"},"nombre":{"type":"string","title":"Nombre","description":"Nombre del proyecto"},"descripcion":{"type":"string","title":"Descripcion","description":"Descripción del proyecto"},"estado":{"type":"string","title":"Estado","description":"Estado actual del proyecto"},"fecha_inicio":{"type":"string","title":"Fecha Inicio","description":"Fecha de inicio del proyecto"},"fecha_fin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Fecha Fin","description":"Fecha de finalización del proyecto"},"materiales":{"items":{"type":"string"},"type":"array","title":"Materiales","description":"Lista de materiales asociados al proyecto","default":[]},"equipos":{"items":{"$ref":"#/components/schemas/Equipo"},"type":"array","title":"Equipos","description":"Lista de equipos asignados al proyecto","default":[]}},"type":"object","required":["nombre","descripcion","estado","fecha_inicio"],"title":"Proyecto"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}